\name{dag_bcell}
\alias{dag_bcell}
\title{DAG representation of the B cell receptor signaling pathway }
\description{
The B cell receptor signaling pathway (BCR) converted to a DAG. In case of nodes with multiple genes, the most differentially expressed
between groups of patients defined in y1 (acute lymphocytic leukemia patients
 with  BCR/ABL gene rearrangement) and y2 (acute lymphocytic leukemia patients
 with  BCR/ABL gene rearrangement) matrices.

Chiaretti S., Li X., Gentleman R., Vitale A., Wang K.S., Mandelli F., Fo R., Ritz J. (2005) Gene expression
profiles of B-lineage adult acute lymphocytic leukemia reveal genetic patterns that identify lineage derivation
and distinct mechanisms of transformation. Clin Cancer Res, 11:7209-7219.
\url{http://clincancerres.aacrjournals.org/content/11/20/7209.long}

Massa, M.S., Chiogna, M., Romualdi, C. (2010). Gene set analysis exploiting the topology of a pathway. BMC Systems Biology, 4:121
\url{http://www.biomedcentral.com/1752-0509/4/121}
}
