\name{print}
\alias{print.clique.var.test}
\alias{print.clique.mean.test}
\alias{print.pathway.var.test}
\alias{print.pathway.mean.test}
\title{Print results of statistical tests.}
\description{
This functions prints a summary of the results of each statistical test.
}
\usage{
\method{print}{clique.var.test}(x, \dots)
\method{print}{clique.mean.test}(x, \dots)
\method{print}{pathway.var.test}(x, \dots)
\method{print}{pathway.mean.test}(x, \dots)
}
\arguments{
\item{x}{the result of a statistical test}
\item{\dots}{ignored}
}
\author{M. Sofia Massa, Gabriele Sales}
