\name{y1}
\alias{y1}
\title{Gene expression signatures in acute lymphocytic leukemia for BCR signaling pathway}
\description{
Several distinct genetic mechanisms lead to acute lymphocytic leukemia (ALL) malignant transformations deriving 
from distinct lymphoid precursor cells that have been committed to either T-lineage or B-lineage differentiation.
Chromosome translocations and molecular rearrangements are common events in B-lineage ALL and reflect distinct 
mechanisms of transformation. The B cell Receptor (BCR/ABL) gene rearrangement occurs in about 25\% of cases in
adult ALL, and much less frequently in pediatric ALL. 
\code{y1} dataset contains gene expression signatures of BCR signaling pathway (composed by 35 gene products, n.columns = 35)
in acute lymphocytic leukemia (ALL) cells associated with presence of BCR/ABL gene rearrangement (n. rows = 37).
Expression values are normalized according to \code{rma} and \code{quantile} normalization.

Chiaretti S., Li X., Gentleman R., Vitale A., Wang K.S., Mandelli F., Fo R., Ritz J. (2005) Gene expression 
profiles of B-lineage adult acute lymphocytic leukemia reveal genetic patterns that identify lineage derivation 
and distinct mechanisms of transformation. Clin Cancer Res, 11:7209-7219.
\url{http://clincancerres.aacrjournals.org/content/11/20/7209.long}

Massa, M.S., Chiogna, M., Romualdi, C. (2010). Gene set analysis exploiting the topology of a pathway. BMC Systems Biology, 4:121
\url{http://www.biomedcentral.com/1752-0509/4/121}
}
